break up gamess, hide registration
