+++ qcengine/programs/__init__.py
@@ -66,6 +66,6 @@ register_program(MolproExecutor())
 register_program(DFTD3Executor())
 register_program(TeraChemExecutor())
 register_program(MP2DExecutor())
-register_program(GAMESSExecutor())
-register_program(NWChemExecutor())
-register_program(CFOURExecutor())
+#register_program(GAMESSExecutor())
+#register_program(NWChemExecutor())
+#register_program(CFOURExecutor())
